requirements.txt
mysql-connector-python
bcrypt
PyQt6
